requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
langgraph==0.2.45
langchain-aws==0.2.6
langchain-core==0.3.15
boto3==1.35.36
pandas==2.2.3
python-dotenv==1.0.1
python-multipart==0.0.12
structlog==24.4.0
numpy==1.26.4